Enable handler auto-loading
No need to load all the handlers by hand, a list of all enabled handlers
is created and they are automagically loaded.

This requires that each handler is in its own file, so had to move
template and gettmp to their own files.

This, again, requires inter-handler communications, which now can be
done by using bot.handlers['name']